+++ src/lsp.rs
@@ -5,9 +5,9 @@ use std::io::{BufRead, BufWriter, Write};
 use std::path::PathBuf;
 
 const COOL_KEYWORDS: &[&str] = &[
-    "def", "extern", "class", "struct", "packed", "union", "if", "elif", "else", "while", "for", "in", "not", "and",
-    "or", "return", "break", "continue", "pass", "import", "from", "as", "try", "except", "finally", "raise", "with",
-    "lambda", "assert", "global", "nonlocal", "True", "False", "None",
+    "def", "data", "extern", "class", "struct", "packed", "union", "if", "elif", "else", "while", "for", "in", "not",
+    "and", "or", "return", "break", "continue", "pass", "import", "from", "as", "try", "except", "finally", "raise",
+    "with", "lambda", "assert", "global", "nonlocal", "True", "False", "None",
 ];
 
 const COOL_BUILTINS: &[&str] = &[